Task 18 Lesson

+++ Module5-server/src/service/articleService.js
@@ -1,5 +1,6 @@
 // const article = require("../database/models/article");
 const articlesRepository = require("../repository/articlesRepository");
+const NotImplementedError = require("../infrastucture/errors/NotemplementedError");
 
 module.exports = {
   getAllArticles: async () => {
@@ -10,21 +11,34 @@ module.exports = {
     const article = await articlesRepository.findArticle(articleId);
     return article;
   },
-  creatNewArticle: async (user, articleData) => {
+  creatNewArticle: async (articleData) => {
     const newArticle = await articlesRepository.createArticle(articleData);
     return newArticle;
   },
-  updateArticle: async (articleId, data) => {
+  updateArticle: async (articleId, data, userId) => {
+    const article = await articlesRepository.findUserArticle(articleId, userId);
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
+
     const updatedArticle = await articlesRepository.updateArticle(
       articleId,
       data
     );
     return updatedArticle;
   },
-  removeArticle: async (articleId) => {
+  removeArticle: async (articleId, userId) => {
+    const article = await articlesRepository.findUserArticle(articleId, userId);
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
     await articlesRepository.removeArticle(articleId);
   },
-  removeArticleForce: async (articleId) => {
+  removeArticleForce: async (articleId, userId) => {
+    const article = await articlesRepository.findUserArticle(articleId, userId);
+    if (!article) {
+      throw new NotImplementedError("Article not found");
+    }
     await articlesRepository.removeArticleForse(articleId);
   },
 };
